refactor: rename filter components and update import paths for consistency
